backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.124.2
fonttools==4.61.0
h11==0.16.0
idna==3.11
joblib==1.5.2
jpype1==1.6.0
kiwisolver==1.4.9
konlpy==0.6.0
lxml==6.0.2
matplotlib==3.10.8
newsapi-python==0.2.7
nltk==3.9.2
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
wordcloud==1.9.4

# Authentication & Database
bcrypt==4.0.1
cryptography==46.0.3
passlib==1.7.4
PyMySQL==1.1.2
python-jose==3.5.0
python-multipart==0.0.20
SQLAlchemy==2.0.45
email-validator==2.3.0
dnspython==2.8.0